Moved the prefixed monolog file to a folder
